Cleanup: Move most classes to their own module.
