+++ ezoutlet/parser.py
@@ -1,7 +1,13 @@
 # Copyright (C) 2015 Schweitzer Engineering Laboratories, Inc.
 # This software may be modified and distributed under the terms
 # of the MIT license.  See the LICENSE file for details.
+
+from __future__ import absolute_import
+from __future__ import print_function
+from __future__ import unicode_literals
+
 import argparse
+import sys
 
 from . import constants
 
@@ -31,3 +37,14 @@ class Parser(object):
 
     def parse_args(self, argv):
         return self._parser.parse_args(argv[1:])
+
+
+def print_help():
+    print(static_parser.get_help(), file=sys.stderr)
+
+
+static_parser = Parser()
+
+
+def print_usage():
+    print(static_parser.get_usage(), file=sys.stderr)
